// checkout-service/src/events/payment-queue.interface.ts
export interface PaymentOrderMessage {
  orderId: string;
  userId: string;
  amount: number;
  items: {
    productId: string;
    quantity: number;
    price: number;
  }[];
  paymentMethod: string;
  description?: string;
  createdAt?: Date;
  metadata?: {
    service: string;
    timestamp: string;
  };
}